+++ test/fixtures/anonlib.js
@@ -7,9 +7,10 @@
  */
 
 'use strict';
-/*
-define(function (require, exports, module) {
-  var dep1 = require('dep1'),
-      dep2 = require('dep2'),
-      dep3 = require('dep3');
-});*/
+
+define(function () {
+  return {
+    foo: 'foo',
+    bar: 'bar'
+  };
+});